requirements.txt
click==8.0.1
greenlet==1.1.0
psycopg2-binary==2.8.6
PyYAML==5.4.1
